backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.36
click==8.1.7
